+++ app/src/impl/load_images.hpp
@@ -1,10 +1,11 @@
 #pragma once
 
+#include <impl/Float.hpp>
 #include <string>
 #include <vector>
 
 namespace impl {
 
-auto load_images(std::string const& name) noexcept -> std::vector<std::vector<double>>;
+auto load_images(std::string const& name) noexcept -> std::vector<std::vector<Float>>;
 
 }  // namespace impl
